Move AccountMenu to account sub directory
